server/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
python-dotenv
flask-cors